worked on payment
